requirements.txt

bcrypt==3.1.7
boto==2.49.0
boto3==1.14.31
botocore==1.17.31
Django==3.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.8
django-environ==0.4.5
django-filter==2.3.0
django-nine==0.2.3
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
elasticsearch==7.8.1
elasticsearch-dsl==7.2.1
oauth==1.0.1
PyJWT==1.7.1
s3transfer==0.3.3
SecretStorage==2.3.1
sendgrid==6.4.3
stripe==2.49.0
virtualenv==20.0.26
mysqlclient==2.0.1
